Minor changes - suggest add comment re small effect
In the Fitting and interpreting a multiple linear regression model with an interaction exercise, is it worth making a comment to say that some of the effect sizes although significant are v. small and sample size is big so likely you have good power to detect small effects that may not be interesting or important (or only important when considered over someone's whole life).

ie. the coefficient for the interaction term is 0.02 and your exploratory visualisation suggests v little observable evidence of a sizeable interaction between age and sex.